test(testing): разделены оставшиеся крупные test suites
article screen renderer, background refresh coordinator, remote sync reload, source management state и settings presentation tests разнесены по behavior-focused файлам
shared doubles и helpers вынесены в отдельные test support файлы
