+++ README.md
@@ -544,7 +544,7 @@
 - [x] `Feed Normalization Rule Split`: разнести `FeedNormalizationService` на metadata normalization, entry normalization / identity input, text & HTML scalar cleanup, URL normalization и icon URL policy, чтобы будущие unit tests normalizer проверяли правила без прохода через один большой static helper;
 - [x] `Large Test Suite Boundary Split`: до эпика `Testing` разделить оставшиеся крупные suites (`ArticlesScreenControllerTests`, `ArticlesScreenStateTests`, `SourceManagementScreenControllerTests`, `SettingsScreenControllerTests`, `FeedRefreshServiceTests`) на behavior-focused files, чтобы новые тесты добавлялись рядом с соответствующим flow;
 - [x] `Source Management Screen Controller Flow Split`: после partition state/view/service разнести `SourceManagementScreenController` на focused controller extensions / helpers для launch routing, add-feed preview/save, move-source submit, create-folder submit и status mapping, чтобы root controller остался координатором `SourceManagementScreenState`, dependencies и presented scenario flow;
-- [ ] `Residual Large Test Suite Split`: проверить и физически разделить оставшиеся крупные focused suites (`ArticleScreenContentRendererTests`, `BackgroundRefreshExecutionCoordinatorTests`, `AppDependenciesRemoteSyncReloadTests`, `SourceManagementScreenStateTests`, `SettingsScreenPresentationTests`) на подфайлы по behavior groups / shared doubles, если это можно сделать без изменения тестового смысла.
+- [x] `Residual Large Test Suite Split`: проверить и физически разделить оставшиеся крупные focused suites (`ArticleScreenContentRendererTests`, `BackgroundRefreshExecutionCoordinatorTests`, `AppDependenciesRemoteSyncReloadTests`, `SourceManagementScreenStateTests`, `SettingsScreenPresentationTests`) на подфайлы по behavior groups / shared doubles, если это можно сделать без изменения тестового смысла.
 
 #### Testing
 - [ ] unit tests для normalizer;
