+++ RSSReaderTests/ArticleScreen/ArticleScreenContentHeaderTests.swift
@@ -1,0 +1,57 @@
+import Foundation
+import Testing
+import UIKit
+@testable import RSSReader
+
+@Suite("Article Screen / Content Rendering / Header")
+@MainActor
+struct ArticleScreenContentHeaderTests {
+    @Test
+    func articleScreenContentHeaderFormatsFieldsInPublishedTitleAuthorFeedOrder() {
+        let publishedAt = Date(timeIntervalSince1970: 1_700_000_000)
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                feedTitle: "THECODE.MEDIA",
+                author: "Юлия Зубарева",
+                publishedAt: publishedAt
+            )
+        )
+
+        #expect(content.header.publishedAtText == ArticleScreenDateFormatter.string(from: publishedAt))
+        #expect(content.header.title == "Article")
+        #expect(content.header.author == "Юлия Зубарева")
+        #expect(content.header.feedTitle == "THECODE.MEDIA")
+    }
+
+    @Test
+    func articleScreenContentHeaderHidesBlankMetadataAndFallsBackForBlankTitle() {
+        let content = ArticleScreenContentState(
+            article: makeReaderArticleDTO(
+                feedTitle: "   ",
+                title: "   ",
+                author: " \n ",
+                publishedAt: nil
+            )
+        )
+
+        #expect(content.header.publishedAtText == nil)
+        #expect(content.header.title == "Untitled Article")
+        #expect(content.header.author == nil)
+        #expect(content.header.feedTitle == nil)
+    }
+
+    @Test
+    func articleScreenStateUsesExistingRenderingPriorityForBodyContent() {
+        var state = ArticleScreenState()
+        let article = makeReaderArticleDTO(
+            summary: "Summary copy",
+            contentHTML: "<p>HTML body</p>",
+            contentText: "Longer content text"
+        )
+
+        state.applyLoadedArticle(article)
+
+        #expect(state.derivedViewState().content?.body.blocks == [.paragraph(.plainText("HTML body"))])
+        #expect(state.derivedViewState().content?.body.source == .contentHTML)
+    }
+}
