+++ RSSReaderTests/ArticleScreen/ArticleScreenContentRendererTestSupport.swift
@@ -1,0 +1,12 @@
+import Foundation
+import Testing
+import UIKit
+@testable import RSSReader
+
+func makeTestImage() -> UIImage {
+        let renderer = UIGraphicsImageRenderer(size: CGSize(width: 1, height: 1))
+        return renderer.image { context in
+            UIColor.red.setFill()
+            context.fill(CGRect(x: 0, y: 0, width: 1, height: 1))
+        }
+}
