+++ RSSReaderTests/Settings/SettingsScreenPresentationContractTests.swift
@@ -2,9 +2,9 @@ import Foundation
 import Testing
 @testable import RSSReader
 
-@Suite("Settings Screen / Presentation")
+@Suite("Settings Screen / Presentation / Contract")
 @MainActor
-struct SettingsScreenPresentationTests {
+struct SettingsScreenPresentationContractTests {
     @Test
     func settingsScreenPresentationBuilderBuildsSectionedContractFromSettingsSnapshot() {
         let snapshot = AppSettingsSnapshot(
@@ -251,94 +251,6 @@ struct SettingsScreenPresentationTests {
     }
 
     @Test
-    func settingsScreenPresentationBuilderEnablesArticleImageCacheResetWhenCacheExists() throws {
-        let sections = SettingsScreenPresentationBuilder.buildSections(
-            from: SettingsScreenInputBuilder.build(from: AppSettingsSnapshot()),
-            hasArticleImageCache: true
-        )
-        let storageSection = try #require(sections.first(where: { $0.id == .storage }))
-
-        #expect(
-            storageSection.items == [
-                .button(
-                    SettingsButtonItemPresentation(
-                        id: .purgeArchivedArticles,
-                        title: "Clear Archived Articles",
-                        subtitle: "Remove archived articles except starred ones from this device and iCloud.",
-                        systemImage: "archivebox",
-                        role: .destructive,
-                        isEnabled: false
-                    )
-                ),
-                .button(
-                    SettingsButtonItemPresentation(
-                        id: .clearArticleImageCache,
-                        title: "Clear Article Image Cache",
-                        subtitle: "Remove article images saved on this device.",
-                        systemImage: "photo.stack",
-                        role: .destructive,
-                        isEnabled: true
-                    )
-                ),
-                .button(
-                    SettingsButtonItemPresentation(
-                        id: .clearSourceIconCache,
-                        title: "Clear Source Icon Cache",
-                        subtitle: "Remove feed icons saved on this device.",
-                        systemImage: "newspaper",
-                        role: .destructive,
-                        isEnabled: false
-                    )
-                )
-            ]
-        )
-    }
-
-    @Test
-    func settingsScreenPresentationBuilderEnablesSourceIconCacheResetWhenCacheExists() throws {
-        let sections = SettingsScreenPresentationBuilder.buildSections(
-            from: SettingsScreenInputBuilder.build(from: AppSettingsSnapshot()),
-            hasSourceIconCache: true
-        )
-        let storageSection = try #require(sections.first(where: { $0.id == .storage }))
-
-        #expect(
-            storageSection.items.last == .button(
-                SettingsButtonItemPresentation(
-                    id: .clearSourceIconCache,
-                    title: "Clear Source Icon Cache",
-                    subtitle: "Remove feed icons saved on this device.",
-                    systemImage: "newspaper",
-                    role: .destructive,
-                    isEnabled: true
-                )
-            )
-        )
-    }
-
-    @Test
-    func settingsScreenPresentationBuilderEnablesArchivedArticlePurgeWhenArchivedArticlesExist() throws {
-        let sections = SettingsScreenPresentationBuilder.buildSections(
-            from: SettingsScreenInputBuilder.build(from: AppSettingsSnapshot()),
-            hasArchivedArticles: true
-        )
-        let storageSection = try #require(sections.first(where: { $0.id == .storage }))
-
-        #expect(
-            storageSection.items.first == .button(
-                SettingsButtonItemPresentation(
-                    id: .purgeArchivedArticles,
-                    title: "Clear Archived Articles",
-                    subtitle: "Remove archived articles except starred ones from this device and iCloud.",
-                    systemImage: "archivebox",
-                    role: .destructive,
-                    isEnabled: true
-                )
-            )
-        )
-    }
-
-    @Test
     func settingsScreenPresentationBuilderIncludesAllSupportedItemTypes() {
         let sections = SettingsScreenPresentationBuilder.buildSections(
             from: SettingsScreenInputBuilder.build(from: AppSettingsSnapshot())
@@ -396,129 +308,5 @@ struct SettingsScreenPresentationTests {
         #expect(input.iCloudSyncStatus == .syncing)
         #expect(input.syncStatusPresentation == .syncing)
         #expect(input.interfaceThemeMode == .black)
-    }
-
-    @Test
-    func settingsScreenPresentationBuilderShowsAppleIDSignInGuidanceWhenNoICloudAccountIsAvailable() throws {
-        let input = SettingsScreenInput(
-            useiCloudSync: true,
-            iCloudSyncStatus: .statusUnavailable,
-            syncStatusPresentation: .noAccount
-        )
-
-        let syncSection = try #require(
-            SettingsScreenPresentationBuilder.buildSections(from: input).first(where: { $0.id == .updatesAndSync })
-        )
-        let statusRow = try #require(syncSection.items.last)
-
-        #expect(
-            statusRow == .statusRow(
-                SettingsStatusRowItemPresentation(
-                    id: .iCloudSyncStatus,
-                    title: "Current Status",
-                    subtitle: "Sign in to iCloud with the Apple ID used on this device to enable sync.",
-                    valueTitle: "Sign In Required"
-                )
-            )
-        )
-        #expect(syncSection.footer?.contains("iCloud sync uses the Apple ID signed in on this device.") == true)
-    }
-
-    @Test
-    func settingsScreenPresentationBuilderShowsSpecificAccountProblemCopyForRestrictedAndTemporaryCases() throws {
-        let restrictedInput = SettingsScreenInput(
-            useiCloudSync: true,
-            iCloudSyncStatus: .statusUnavailable,
-            syncStatusPresentation: .restricted
-        )
-        let temporarilyUnavailableInput = SettingsScreenInput(
-            useiCloudSync: true,
-            iCloudSyncStatus: .statusUnavailable,
-            syncStatusPresentation: .temporarilyUnavailable
-        )
-        let couldNotDetermineInput = SettingsScreenInput(
-            useiCloudSync: true,
-            iCloudSyncStatus: .statusUnavailable,
-            syncStatusPresentation: .couldNotDetermine
-        )
-
-        let restrictedSection = try #require(
-            SettingsScreenPresentationBuilder.buildSections(from: restrictedInput).first(where: { $0.id == .updatesAndSync })
-        )
-        let temporarilyUnavailableSection = try #require(
-            SettingsScreenPresentationBuilder.buildSections(from: temporarilyUnavailableInput).first(where: { $0.id == .updatesAndSync })
-        )
-        let couldNotDetermineSection = try #require(
-            SettingsScreenPresentationBuilder.buildSections(from: couldNotDetermineInput).first(where: { $0.id == .updatesAndSync })
-        )
-
-        #expect(
-            restrictedSection.items.last == .statusRow(
-                SettingsStatusRowItemPresentation(
-                    id: .iCloudSyncStatus,
-                    title: "Current Status",
-                    subtitle: "This device cannot use iCloud right now because account changes or CloudKit access are restricted.",
-                    valueTitle: "Restricted"
-                )
-            )
-        )
-        #expect(
-            temporarilyUnavailableSection.items.last == .statusRow(
-                SettingsStatusRowItemPresentation(
-                    id: .iCloudSyncStatus,
-                    title: "Current Status",
-                    subtitle: "The current iCloud account is temporarily unavailable. Try again later.",
-                    valueTitle: "Temporarily Unavailable"
-                )
-            )
-        )
-        #expect(
-            couldNotDetermineSection.items.last == .statusRow(
-                SettingsStatusRowItemPresentation(
-                    id: .iCloudSyncStatus,
-                    title: "Current Status",
-                    subtitle: "The app could not determine the current iCloud account status. Check the device Apple ID and iCloud availability, then try again.",
-                    valueTitle: "Account Unavailable"
-                )
-            )
-        )
-    }
-
-    @Test
-    func settingsScreenPresentationBuilderExplainsLocalOnlyFallbackWhenSyncPreferenceIsSavedButBootstrapStayedLocal() throws {
-        let input = SettingsScreenInput(
-            useiCloudSync: true,
-            iCloudSyncStatus: .statusUnavailable,
-            syncStatusPresentation: .temporarilyUnavailable,
-            isUsingLocalOnlySyncFallbackForCurrentLaunch: true
-        )
-
-        let syncSection = try #require(
-            SettingsScreenPresentationBuilder.buildSections(from: input).first(where: { $0.id == .updatesAndSync })
-        )
-
-        #expect(
-            syncSection.items.dropFirst().first == .toggle(
-                SettingsToggleItemPresentation(
-                    id: .useICloudSync,
-                    title: "Enable iCloud Sync",
-                    subtitle: "Saved for the next launch. This session keeps using local data because the current iCloud account is temporarily unavailable.",
-                    isOn: true
-                )
-            )
-        )
-        #expect(
-            syncSection.items.last == .statusRow(
-                SettingsStatusRowItemPresentation(
-                    id: .iCloudSyncStatus,
-                    title: "Current Status",
-                    subtitle: "Sync is enabled, but this launch cannot use iCloud because the current account is temporarily unavailable. Relaunch after iCloud becomes available.",
-                    valueTitle: "Temporarily Unavailable"
-                )
-            )
-        )
-        #expect(
-            syncSection.footer?.contains("Sync will try again on the next launch when iCloud is available.") == true
-        )
     }
 }
