+++ RSSReaderTests/SourceManagement/SourceManagementAddFeedScreenStateTests.swift
@@ -2,20 +2,9 @@ import Foundation
 import Testing
 @testable import RSSReader
 
-@Suite("Source Management / Screen State")
+@Suite("Source Management / Screen State / Add Feed")
 @MainActor
-struct SourceManagementScreenStateTests {
-    @Test
-    func sourceManagementScreenStateBuildsSeparatedEntrySections() {
-        let state = SourceManagementScreenState.makePreviewFixture()
-        let viewState = state.derivedViewState()
-
-        #expect(viewState.summary.title == "Manage sources and folders.")
-        #expect(viewState.sections.map(\.id) == [.startNew, .organizeExisting])
-        #expect(viewState.sections.first?.items.map(\.id) == [.addFeed, .createFolder])
-        #expect(viewState.sections.last?.items.map(\.id) == [.moveSource])
-    }
-
+struct SourceManagementAddFeedScreenStateTests {
     @Test
     func sourceManagementScreenStateBuildsAddFeedPresentationWithPreviewAndSingleSaveState() {
         var state = SourceManagementScreenState.makePreviewFixture()
@@ -366,180 +355,5 @@ struct SourceManagementScreenStateTests {
         #expect(retryLoadingDestination.isLoadingPreview)
         #expect(retryLoadingDestination.preview == nil)
         #expect(retryLoadingDestination.status == nil)
-    }
-
-    @Test
-    func sourceManagementScreenStateBuildsCreateFolderPresentationWithValidationAndPlacement() {
-        var state = SourceManagementScreenState.makePreviewFixture()
-        state.applyCreateFolderContext(
-            folders: [
-                SourceManagementFolderSummary(
-                    id: UUID(uuidString: "33333333-3333-3333-3333-333333333333")!,
-                    name: "News",
-                    sortOrder: 0,
-                    feedCount: 5
-                ),
-                SourceManagementFolderSummary(
-                    id: UUID(uuidString: "44444444-4444-4444-4444-444444444444")!,
-                    name: "Tech",
-                    sortOrder: 1,
-                    feedCount: 3
-                )
-            ]
-        )
-        state.presentScenario(.createFolder)
-
-        guard case .createFolder(let initialDestination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation")
-            return
-        }
-
-        #expect(initialDestination.existingFolders.map(\.name) == ["News", "Tech"])
-        #expect(initialDestination.placementDescription == "This folder will be added after 2 existing folders.")
-        #expect(initialDestination.isPrimaryActionEnabled == false)
-
-        state.updateCreateFolderNameInput("News")
-
-        guard case .createFolder(let duplicateDestination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation after duplicate input")
-            return
-        }
-
-        #expect(duplicateDestination.validationMessage == "A folder with this name already exists.")
-        #expect(duplicateDestination.isPrimaryActionEnabled == false)
-
-        state.updateCreateFolderNameInput(" tEcH ")
-
-        guard case .createFolder(let caseDuplicateDestination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation after case-insensitive duplicate input")
-            return
-        }
-
-        #expect(caseDuplicateDestination.validationMessage == "A folder with this name already exists.")
-        #expect(caseDuplicateDestination.isPrimaryActionEnabled == false)
-
-        state.updateCreateFolderNameInput("Research")
-
-        guard case .createFolder(let validDestination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected create-folder destination presentation after valid input")
-            return
-        }
-
-        #expect(validDestination.validationMessage == nil)
-        #expect(validDestination.isPrimaryActionEnabled)
-    }
-
-    @Test
-    func sourceManagementScreenStateBuildsMoveSourcePresentationWithFeedAndPlacementSelection() {
-        var state = SourceManagementScreenState.makePreviewFixture()
-        let newsFolderID = UUID(uuidString: "55555555-5555-5555-5555-555555555555")!
-        let techFolderID = UUID(uuidString: "66666666-6666-6666-6666-666666666666")!
-        state.applyMoveSourceContext(
-            feeds: [
-                SourceManagementFeedSummary(
-                    id: UUID(uuidString: "77777777-7777-7777-7777-777777777777")!,
-                    url: "https://example.com/apple.xml",
-                    title: "Apple Feed",
-                    folderID: newsFolderID,
-                    folderName: "News"
-                ),
-                SourceManagementFeedSummary(
-                    id: UUID(uuidString: "88888888-8888-8888-8888-888888888888")!,
-                    url: "https://example.com/beta.xml",
-                    title: "Beta Feed",
-                    folderID: nil,
-                    folderName: nil
-                )
-            ],
-            folders: [
-                SourceManagementFolderSummary(
-                    id: newsFolderID,
-                    name: "News",
-                    sortOrder: 0,
-                    feedCount: 2
-                ),
-                SourceManagementFolderSummary(
-                    id: techFolderID,
-                    name: "Tech",
-                    sortOrder: 1,
-                    feedCount: 6
-                )
-            ]
-        )
-        state.presentScenario(.moveSource)
-
-        guard case .moveSource(let initialDestination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected move-source destination presentation")
-            return
-        }
-
-        #expect(initialDestination.feeds.map(\.title) == ["Apple Feed", "Beta Feed"])
-        #expect(initialDestination.feeds.first?.isSelected == true)
-        #expect(initialDestination.placementOptions.map(\.title) == ["Ungrouped", "News", "Tech"])
-        #expect(initialDestination.isPrimaryActionEnabled == false)
-
-        state.selectMoveSourcePlacement(.folder(techFolderID))
-
-        guard case .moveSource(let changedDestination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected move-source destination presentation after placement change")
-            return
-        }
-
-        #expect(changedDestination.isPrimaryActionEnabled)
-        #expect(changedDestination.placementOptions.last?.isSelected == true)
-    }
-
-    @Test
-    func sourceManagementScreenStatePreselectsRequestedFeedForMoveSourceContext() {
-        var state = SourceManagementScreenState.makePreviewFixture()
-        let newsFolderID = UUID(uuidString: "55555555-5555-5555-5555-555555555555")!
-        let techFolderID = UUID(uuidString: "66666666-6666-6666-6666-666666666666")!
-        let appleFeedID = UUID(uuidString: "77777777-7777-7777-7777-777777777777")!
-        let betaFeedID = UUID(uuidString: "88888888-8888-8888-8888-888888888888")!
-
-        state.applyMoveSourceContext(
-            feeds: [
-                SourceManagementFeedSummary(
-                    id: appleFeedID,
-                    url: "https://example.com/apple.xml",
-                    title: "Apple Feed",
-                    folderID: newsFolderID,
-                    folderName: "News"
-                ),
-                SourceManagementFeedSummary(
-                    id: betaFeedID,
-                    url: "https://example.com/beta.xml",
-                    title: "Beta Feed",
-                    folderID: techFolderID,
-                    folderName: "Tech"
-                )
-            ],
-            folders: [
-                SourceManagementFolderSummary(
-                    id: newsFolderID,
-                    name: "News",
-                    sortOrder: 0,
-                    feedCount: 2
-                ),
-                SourceManagementFolderSummary(
-                    id: techFolderID,
-                    name: "Tech",
-                    sortOrder: 1,
-                    feedCount: 6
-                )
-            ],
-            selectedFeedID: betaFeedID
-        )
-        state.presentScenario(.moveSource)
-
-        guard case .moveSource(let destination)? = state.derivedViewState().presentedDestination else {
-            Issue.record("Expected move-source destination presentation")
-            return
-        }
-
-        #expect(destination.feeds.first(where: { $0.id == appleFeedID })?.isSelected == false)
-        #expect(destination.feeds.first(where: { $0.id == betaFeedID })?.isSelected == true)
-        #expect(destination.placementOptions.first(where: { $0.title == "Tech" })?.isSelected == true)
-        #expect(destination.isPrimaryActionEnabled == false)
     }
 }
